[octopus] fixed linux compil

+++ src/godoctopus/src/Controller.cpp
@@ -121,7 +121,7 @@ void Controller::_process(double delta)
 
 		if(_state && _state->getStepApplied() % 128 == 0)
 		{
-			emit_signal("state_dump", (long long)_state->getStepApplied(), hashState(*_state));
+			emit_signal("state_dump", (int64_t)_state->getStepApplied(), hashState(*_state));
 		}
 	}
 }
